Use insserv; preserve flud.conf
Prefer insserv over update-rc.d for installing flud as a service.
Flud is dependent on the network, and also some basic services which we
roll up as a dependency on syslog.
Wherever possible, preserve the existing flud.conf configuration, which
means copying from host on remote install, and not removing it on
uninstall.