Initial use of jQuery.  Implements "startValve."
Begin to use jQuery, and ajax calls to server.
Convert "advance", and "stop" to use this method.
Implement "startValve" to start at specified valve, which is more
convenient than advancing through valves to the desired one.

Note that this implementation is still rough.  After a valve-changing
call this "<- ON" marker is not updated, and the status/result area is
really poorly done.  This all still needs a bunch of work.